switch to node test runner in typeorm example
